Move error parsing code from utils.rs to error.rs
